Address PR review feedback
- Split embeddings.py into analysis/embeddings/ subpackage
  - client.py: API interaction and EmbeddingResult
  - text_builders.py: build_analysis_text, build_summary_text, etc.
  - chunking.py: chunk_text and estimate_tokens
  - operations.py: embed_conversation_full, check_embedding_status

- Extract RAG logic from cli.py to analysis/rag.py
  - Create RAGAnswerer class with answer_question method
  - Simplify CLI to thin wrapper around RAGAnswerer

- Fix cache.py implicit side effects
  - Add update_embeddings=False parameter to save() method
  - Embedding updates now require explicit opt-in

- Fix exception handling in cache.py
  - Replace broad 'except Exception' with specific exceptions
  - Catch (ImportError, RuntimeError, OSError) for expected failures

- Simplify migrations
  - Combine 008_embeddings.py and 009_embedding_types.py
  - Feature never merged, so single clean schema is cleaner

- Relax numpy version constraint
  - Changed from >=2.4.4 to >=2.0,<3.0 for broader compatibility
