+++ pyproject.toml
@@ -9,7 +9,7 @@ authors = [{ name = "JP Shackelford" }]
 dependencies = [
     "click>=8.0",
     "httpx>=0.27",
-    "numpy>=2.4.4",
+    "numpy>=2.0,<3.0",
     "openhands-sdk>=1.16.1",
     "pyyaml>=6.0",
     "rich>=13.0",
